refactored Popup
